change ErrorHandler for Jetty 12 API
