+++ core/src/main/java/io/confluent/rest/ApplicationServer.java
@@ -20,7 +20,7 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
-import io.confluent.rest.errorhandlers.NoJettyDefaultStackTraceErrorHandler;
+import io.confluent.rest.errorhandlers.StackTraceErrorHandler;
 import java.lang.management.ManagementFactory;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -217,9 +217,7 @@ public final class ApplicationServer<T extends RestConfig> extends Server {
   @Override
   protected final void doStart() throws Exception {
     // set the default error handler
-    if (config.getSuppressStackTraceInResponse()) {
-      this.setErrorHandler(new NoJettyDefaultStackTraceErrorHandler());
-    }
+    this.setErrorHandler(new StackTraceErrorHandler(config.getSuppressStackTraceInResponse()));
 
     Sequence handlers = new Sequence();
     Sequence wsHandlers = new Sequence();
